+++ src/main/java/io/github/janguenter/bluemap/trophymanager/adapter/bluemap523/TrophyBlockEntityData.java
@@ -1,6 +1,6 @@
 /* SPDX-License-Identifier: MIT */
 
-package io.github.janguenter.bluemap.trophymanager.adapter.bluemap522;
+package io.github.janguenter.bluemap.trophymanager.adapter.bluemap523;
 
 import de.bluecolored.bluemap.core.world.mca.blockentity.MCABlockEntity;
 import de.bluecolored.bluenbt.NBTName;