+++ src/main/java/io/github/janguenter/bluemap/trophymanager/adapter/bluemap523/VariantRendererCatalog.java
@@ -1,6 +1,6 @@
 /* SPDX-License-Identifier: MIT */
 
-package io.github.janguenter.bluemap.trophymanager.adapter.bluemap522;
+package io.github.janguenter.bluemap.trophymanager.adapter.bluemap523;
 
 import de.bluecolored.bluemap.core.map.hires.block.BlockRendererType;
 import de.bluecolored.bluemap.core.resources.pack.resourcepack.ResourcePack;