+++ src/test/java/io/github/janguenter/bluemap/trophymanager/adapter/bluemap523/TrophyRenderPlanTest.java
@@ -1,6 +1,6 @@
 /* SPDX-License-Identifier: MIT */
 
-package io.github.janguenter.bluemap.trophymanager.adapter.bluemap522;
+package io.github.janguenter.bluemap.trophymanager.adapter.bluemap523;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;